+++ CHANGELOG.md
@@ -2,14 +2,42 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.12.0] - 2021-03-13
+
+### Changed
+
+- Adjustments to null safety
+- Function `merge` has been deprecated due to null-safety changes.
+- Updated dependencies:
+    - http: ^0.13.0
+    - json_annotation: ^4.0.0
+- Updated development dependencies:
+    - json_serializable: ^4.0.2
+    - build_runner: ^1.10.1
+    - pedantic: ^1.11.0
+
+## [0.10.0] - 2021-03-10
+
+### Added
+
+- Time FQL functions:
+    - `TimeAdd`, `TimeSubtract`, `TimeDiff`
+
+## [0.9.1] - 2021-02-25
+
+### Changed
+
+- Fixed `Join` FQL function (incorrect JSON name of parameter)
+
 ## [0.9.0] - 2020-12-14
 
 ### Added
 
 - Authentication FQL functions:
-    - Credentials, CurrentIdentity, CurrentToken, HasCurrentToken
+    - `Credentials`, `CurrentIdentity`, `CurrentToken`, `HasCurrentToken`
 - Conversion FQL functions:
-    - ToArray, ToDate, ToDouble, ToInteger, ToMicros, ToMillis, ToNumber, ToObject, ToSeconds, ToString, ToTime
+    - `ToArray`, `ToDate`, `ToDouble`, `ToInteger`, `ToMicros`, `ToMillis`, `ToNumber`, `ToObject`, `ToSeconds`
+      , `ToString`, `ToTime`
 
 ### Changed
 
@@ -125,6 +153,10 @@ All notable changes to this project will be documented in this file.
 
 - `FaunaDBConfig` class
 - `FaunaDBClient` class
+
+[0.10.0]: https://github.com/gavanitrate/faunadb-http-dart/compare/0.9.1...0.10.0
+
+[0.9.1]: https://github.com/gavanitrate/faunadb-http-dart/compare/0.9.0...0.9.1
 
 [0.9.0]: https://github.com/gavanitrate/faunadb-http-dart/compare/0.8.1...0.9.0